Load NHD layers into current session

Load NHD layers into current session

nhd_load(state, dsn, file_ext = NA, approve_all_dl = FALSE, ...)

Arguments

state
character state abbreviation

dsn
character name of a NHD layer

file_ext
character choice of "shp" for spatial data and "dbf" or "gpkg" for non-spatial. optional

approve_all_dl
logical blanket approval to download all missing data

...
arguments passed to sf::st_read

Value

Spatial simple features object or data frame depending on the dsn type and value passed to file_ext

Details

This function will ask the user to approve downloading missing data unless approve_all_dl is set to TRUE.

Examples

# NOT RUN {
dt <- nhd_load(c("RI"), c("NHDWaterbody"))
dt <- nhd_load(c("CT", "RI"), "NHDWaterbody")
dt <- nhd_load(c("CT", "RI"), "NHDWaterbody", quiet = TRUE)
dt <- nhd_load("MI", "NHDFlowline")
dt <- nhd_load("RI", "NHDReachCrossReference")
dt <- nhd_load("RI", "NHDWaterbody", file_ext = "dbf")
dt <- nhd_load(c("RI", "DC"), "NHDWaterbody", file_ext = "gpkg")
# }