Select features clipped by a point buffer around a point

Select features clipped by a point buffer around a point

select_point_overlay(pnt, sp, buffer_dist = 0.05)

Arguments

pnt
geographic point of class sfc

sp
list of sf data frames

buffer_dist
numeric buffer in units of coordinate degrees

Examples

if (FALSE) {
wk <- wikilake::lake_wiki("Gull Lake (Michigan)")
pnt <- sf::st_sfc(sf::st_point(c(wk$Lon, wk$Lat)))
sf::st_crs(pnt) <- 4326
sp <- lapply(c("NHDWaterbody", "NHDFlowLine"),
          function(x) nhd_plus_load(vpu = 4, dsn = x))
names(sp) <- c("NHDWaterbody", "NHDFlowLine")
qry <- select_point_overlay(pnt = pnt, sp = sp, buffer_dist = 0.05)
plot(qry$NHDWaterbody$geometry)

}